+++ src/file/filesystem.h
@@ -35,8 +35,8 @@ struct bd_file_s
     int64_t (*seek)(BD_FILE_H *file, int64_t offset, int32_t origin);
     int64_t (*tell)(BD_FILE_H *file);
     int (*eof)(BD_FILE_H *file);
-    int (*read)(BD_FILE_H *file, uint8_t *buf, int64_t size);
-    int (*write)(BD_FILE_H *file, const uint8_t *buf, int64_t size);
+    int64_t (*read)(BD_FILE_H *file, uint8_t *buf, int64_t size);
+    int64_t (*write)(BD_FILE_H *file, const uint8_t *buf, int64_t size);
 };
 
 /*
